+++ CMakeLists.txt
@@ -7,8 +7,8 @@ set(CMAKE_CXX_EXTENSIONS OFF)
 list(APPEND CMAKE_MODULE_PATH ${CMAKE_BINARY_DIR})
 set(CMAKE_CXX_STANDARD 17)
 
-include_directories(model seedData)
-add_subdirectory(problems)
+include_directories(include)
+add_subdirectory(src/problems)
 
 enable_testing()
 add_subdirectory(tests)
