+++ include/model/DistanceMap.hpp
