+++ include/model/LoyaltyTier.hpp
