--- problems/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_executable(app main.cpp)
-target_include_directories(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
