+++ src/problems/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_executable(app main.cpp)
